Adds Hyper-V serial console log
Introduces the VM serial output log feature to the Hyper-V Nova
 Compute driver.

The VM serial port COM0 is redirected to a named pipe so that a
worker can take the VM output and write it to a file.

The VM serial output is then read from this file whenever the
console log command is issued.

Implements: blueprint hyper-v-console-log

Co-Authored-By: Adelina Tuvenie <atuvenie@cloudbasesolutions.com>

Change-Id: I7d3132c0f934ac3e8d28315f4da766466612a6c0

Conflicts:
	nova/tests/virt/hyperv/test_hypervapi.py
	nova/tests/virt/hyperv/test_vmops.py
	nova/tests/virt/hyperv/test_vmutils.py
	nova/virt/hyperv/vmops.py

Conflicts:
	nova/tests/virt/hyperv/test_vmops.py
	nova/virt/hyperv/vmops.py